Renaming dirs with scripts and data
