+++ scripts/pythonScript/src/log2json/parser.py
