+++ scripts/pythonScript/test-resources/resources-parser/stdout_seed1M1.log
