+++ scripts/pythonScript/test-resources/resources-parser/stdout_seed1_c25.log
